(prog) Working on variables expansion and tokenization
